Bugfix uninitialised vars and conflicting types in simulators after merge

+++ openair1/SIMULATION/NR_PHY/dlsim.c
@@ -63,7 +63,7 @@ RAN_CONTEXT_t RC;
 int32_t uplink_frequency_offset[MAX_NUM_CCs][4];
 
 double cpuf;
-int nfapi_mode=0;
+uint8_t nfapi_mode = 0;
 uint16_t NB_UE_INST = 1;
 
 //Dummy Functions
